Move Utils.doFetchFile to specific package and do some code refactor
