+++ shims/spark320/src/main/scala/com/intel/oap/sql/shims/spark320/Spark320Shims
@@ -22,7 +22,7 @@ import java.io.File
 import org.apache.spark.SparkConf
 import org.apache.spark.shuffle.MigratableResolver
 import org.apache.spark.shuffle.ShuffleHandle
-import org.apache.spark.shuffle.ShuffleUtil
+import org.apache.spark.util.ShimUtils
 import org.apache.spark.shuffle.api.ShuffleExecutorComponents
 import org.apache.spark.shuffle.sort.SortShuffleWriter
 import org.apache.spark.sql.catalyst.plans.physical.Partitioning
@@ -35,7 +35,7 @@ class Spark320Shims extends SparkShims {
 
   override def shuffleBlockResolverWriteAndCommit(shuffleBlockResolver: MigratableResolver,
                                                   shuffleId: Int, mapId: Long, partitionLengths: Array[Long], dataTmp: File): Unit =
-    ShuffleUtl.shuffleBlockResolverWriteAndCommit(shuffleBlockResolver, shuffleId, mapId, partitionLengths, dataTmp)
+    ShimUtils.shuffleBlockResolverWriteAndCommit(shuffleBlockResolver, shuffleId, mapId, partitionLengths, dataTmp)
 
   override def getDatetimeRebaseMode(fileMetaData: FileMetaData, parquetOptions: ParquetOptions):
   SQLConf.LegacyBehaviorPolicy.Value = {
@@ -86,7 +86,7 @@ class Spark320Shims extends SparkShims {
   override def newSortShuffleWriter(resolver: MigratableResolver, shuffleHandle: ShuffleHandle,
                                     mapId: Long, context: TaskContext,
                                     shuffleExecutorComponents: ShuffleExecutorComponents): AnyRef = {
-    ShuffleUtil.newSortShuffleWriter(
+    ShimUtils.newSortShuffleWriter(
       resolver,
       shuffleHandle,
       mapId,
@@ -118,7 +118,7 @@ class Spark320Shims extends SparkShims {
 
   override def doFetchFile(urlString: String, targetDirHandler: File,
                            targetFileName: String, sparkConf: SparkConf): Unit = {
-    Utils.doFetchFile(urlString, targetDirHandler, targetFileName, sparkConf, null)
+    ShimUtils.doFetchFile(urlString, targetDirHandler, targetFileName, sparkConf, null)
   }
 
 //  /**
@@ -173,7 +173,7 @@ class Spark320Shims extends SparkShims {
   /**
     * Only applicable to AQEShuffleReadExec. Otherwise, an exception will be thrown.
     */
-  override def getPartitionSpecsOfCustomShuffleReaderExec(plan: SparkPlan): ShufflePartitionSpec = {
+  override def getPartitionSpecsOfCustomShuffleReaderExec(plan: SparkPlan): Seq[ShufflePartitionSpec] = {
     plan match {
       case plan: AQEShuffleReadExec => plan.partitionSpecs
       case _ => throw new RuntimeException("AQEShuffleReadExec is expected!")
